Seperate Seqflow Config model into seperate file
