+++ instruments/seqflow/src/seqflow/seqflow_models.py
@@ -14,16 +14,6 @@ from pathlib import Path
 DeviceType = Literal["pump", "heat_device", "wait", "stopper"]
 
 
-class SeqFlowConfig(BaseModel, validate_assignment=True):
-    # Use default_factory to satisfy mypy's static type checker
-    save_folder: Path = Field(default_factory=lambda: Path("../../seq_flow_jobs/"))
-    
-    selector_port_map: dict[str, int]
-    instrument_name: Optional[str] = Field(
-        default=None, 
-        description="Optional instrument name."
-    )
-
 class SeqFlowStep(BaseModel):
     """Model representing a single action within a sequence."""
     # --- STEP PARAMETERS ---
